Add files via upload
can_simulator.py -script, which includes two processes simulating CAN communication on ‘vcan0’ and ‘vcan1’.The script sends periodic CAN messages containing ASCII values of "HelloCAN" and sleeps for one second between messages.
central_control_unit.py-This script is to act as a control unit for a node on a simulated CAN bus. It continuously listens for incoming CAN messages and handles them based on their message ID. The script is designed to demonstrate how a node can process and respond differently to emergency messages (with ID ‘0x100’) and regular messages. 
can_visualizer.py- Here it shows to receive CAN messages from the specified CAN channel (‘vcan0’ or ‘vcan1’) and updates a real-time plot based on the received data.
check.py- By thinking of a car it generates messages with different priorities and simulates the handling of these messages by the nodes. The Matplotlib library is used to visualize the CAN message priorities over time. the code simulates a CAN (Controller Area Network) system where messages are sent, received, and prioritized based on their importance. The focus is on how well the Main Control node handles emergency warnings.

check3.py-In the modified code (‘check3.py’), the ‘shared_variable’ is an example of a shared variable. It is a simple counter that is accessed and modified by multiple threads (‘Main Control’, ‘Node 1’, ‘Node 2’) without proper synchronization. This lack of synchronization creates a race condition, as the threads interfere with each other's access and modification of the shared variable.

To address race conditions and ensure proper synchronization, the code uses a lock (‘lock’) to to potentially create a race condition ,protect access to the shared variable. The lock ensures that only one thread at a time can access or modify the shared variable, preventing race conditions and maintaining data consistency.